snapshot 0.2
  - added the game loop inside run() method of the Main class, so that
rendering and updating are perofrmed at a fixed speed.
  - the amount of updates per second (ups) and frames per second (fps)
are printed to console after every iteration of the game loop.
  - moved the render() method from Main to a newly created Renderer
class inside a newly created renderEngine package
  - used lambda expressions to ease the code for collecting input events
(modified the method glfwSetKeyCallback(...) so that parameters of the
invoke(...) method are directly passed into it. The Input class is as of
now deprecated and no longer in use.