snapshot 0.3 - various code corrections
Changed:
  - the glfwSwapBuffers(window.windowHandle) method was incorrectly
placed inside the render() method of the Renderer class in the previous
snapshot. In this snapshot, it is correctly placed in the update()
method of the Main class. It's logical if you think about it... you
should swap the buffers in the update method and not in the render
method

Added:
  - more window hints so as to create the GLFW 3.2 context, specified to
use the core functionality and set the forward compatibility of GLFW to
true --> all deprecated functionalities are deactivated
  - a keycallback instance variable of type GLFWKeyCallback to the
Window class and used it inside the glfwSetKeyCallback(...). The purpose
of this is to be able to free the callback in the main class, which was
previously not possible due to the fact that there was no variable
referring to the created key callback