snapshot 0.4 - rendered a quad to the window using VBOs and VAOs
  - accomplished to draw two black triangles to form a quad on a white
window background:

  - added Mesh class inside renderEngine package. It's main
responsibility is to collect 3D vertices, store them in VRAM and provide
way to acces them through an instance of itself (this instance of Mesh
class will be used by Renderer class to render the vertices referenced
by the mesh instance onto the window)
  
  - added Renderer class, which a) manages window resize events; b)
draws the provided vertices using information provided by a mesh
instance which is passed to it's render(Window window, Mesh mesh)
method.
  
  - set up a window resize callback (dynamically changes the width and
height fields of the Window class if a window is resized), using
glfwSetFramebufferSizeCallback.
  
  - added a temporary runAssertions() method, whose job is just for
debugging purposes - printing out various data to the console, so it can
be seen whether the code behaves as it should. This method will be later
removed and substituted by a whole class, whose role will be similar to
the one of runAssertions() method