snapshot 0.6: added shader support
  - created a new package called shaders, which contains
ShaderProgram.java, fragmentShader.fs and vertexShader.vs
  - ShaderProgram class is responsible for performing different shader
operations - creating them, attaching and linkking them  to the program,
checking the success of shader creation and compilation and providing
methods for binding shaders
  - vertexShader.vs represents the vertex shader
  - fragmentShader.fs represents the fragment shader
  - addedFileUtilities class inside the utils package, which is
responsible for performing various operations regarding file management.
Currently, it only performs reading proper file paths for use with
shaders
  - added init() method to the Renderer class, which creates new shader
objects and links them. This method should be called from the init()
method of the Main class
  - added runAssertions() method to the Renderer class, which checks for
the value of ID. This method should be called by the runAssertions()
method of the Main class.